edmv: escape filenames better
I haven't been bit by it, since filenames rarely include double quotes,
backticks, or backslashes, but it's bugged me for a while that edmv
assumed they were never included.

The bash manpage, under the QUOTING section, says only the dollar sign,
backtick, double-quote, and backslash characters are special within
double quotes. I generally find double-quotes to be the most readable
escaped form, since apostrophes and spaces are common when files are
named after a human creator. The %q format specifier to printf creates
backslash-quoted filenames that can be used as input, and the ${param@Q}
expansion creates single-quoted filenames that can be used as input, but
I haven't found any convenient way to escape a filename that can be used
as double-quoted input. ${param@Q} is a pretty new bash feature (4.4),
and portability is a big part of the motivation for edmv, so avoid using
it.

- give `-r` to read to avoid eating backslashes.
- add a little message to the script if there's nothing to do
- bail out early if no files are given.