Add more bindings for copy/paste using xsel
Define an operator instead of giving a range to a command so that
intra-line selections can be copied and text objects can be used for
selecting.

This is intended as a backup for +clipboard support in vim, since you'll
often want copy/paste to just work _now_ without having to install a
more featureful version of vim or restart it to restore its connection
to the X server.

We need to use systemlist() instead of system() so we can give a list of
lines to append(), since when LF is stored in a Vim
variable/register/etc, it is stored as a NUL, and then translated back
to a real newline under circumstances such as :put. See
https://vim.fandom.com/wiki/Newlines_and_nulls_in_Vim_script