Limit resident memory to 4GB per process
Every once in a while I run a program that accidentally consumes memory
so voraciously that it basically hangs the machine. The OOM killer never
saves me in these situations, so it seems better to kill programs that
try to use tons of memory and make exceptions for the rare ones that
legitimately need that much.