pythonrc: add function to reload a package/prefix
Of course, it's impossible to reliable reload arbitrary packages in
python, but for a limited case where you're modifying a package you know
well and then playing with the changes interactively and the
dependencies are well understood it can be quite convenient to reload
code while keeping some expensive or extensive data structures in
memory.

I suppose in a lot of cases a better alternative approach would be
script up the creation of the objects you want to play would and bind
them to variables in the global scope.