Put short messages into the statusbar
Previously I was treating every message kind of like an email, and was
only displaying the number of messages in the statusbar, but for lots of
little reminders that don't have any associated data it's more
convenient to put more specific text in the statusbar. Let's just add
all the filenames in a particular dir to the statusbar message.