vimrc: improve binding to jump to first quickfix error
I generally don't want to jump to the first error after `:make`, since
the jump can be disorienting and if the error is in another file it'll
destroy the undo buffer for the current file. But if I don't jump to the
first error (that is, if I leave off the ! from :make), it's annoying to
get to the first valid error on my own. `:cc 1` goes to the first error
regardless whether it's valid or associated with a file. `:cn` skips
invalid errors, but trying to go to the second valid error and then
going back to the first is annoying if I don't know where I am in the
quickfix list or if there's only one error and then you get a "No more
items" message.

Saving undo files (:h undo-persistence) would make automatically jumping
to the first error more acceptable, but I think I still might want this
binding anyway.