Only set LANG when needed
For physical machines I want en_US.UTF-8 so that xorg loads compose key
bindings and terminals that enable UTF8 based on LANGUAGE do so. It'd be
nice to use a different locale for LC_TIME to get ISO8601 time formats,
but I can't be bothered to ensure that a Eurpopean locale is generated
on every system, and I'm not sure what'll happen if the set locale
doesn't exist. For more barebones systems like servers, chroots, VMs,
etc, that I'm not running a desktop env on I'm ok with using the system
setting, as long as it's C, POSIX, or en_US.

If the locale is set to something exotic, let's assume that en_US is also
available.