Change keybindings to be more emacs-friendly
I've used C-x as my tmux prefix for a long time but I sometimes wonder
how much I'm missing out on by not having it available as prefix for
emacs-style bindings. So, change to C-q. It's comparably easy to hit and
even if C-q is used in some program, it's probably to quit or something,
and so shouldn't be that bad to pass through tmux.

Get rid of the binding for C-o to go to the previous pane in tmux. Not
sure why I had that; the default binding for it is to go to the next
pane. Note that C-; is bound to last-pane, which is nice.

Add a few more emacs bindings to insert mode in Vim.