vim: override mappings for pasting selection/clipboard
Use the expression pseudo-register to shell out to xsel.

The latest vim package for Arch doesn't include clipboard support, and
using vim's connection to X11 has always been flaky for me, anyway. The
DISPLAY env var might not be set even though X11 is running, and it
seemed like when I detached a tmux session all the vim instances in it
would be unable to paste from the clipboard anymore.

I should have done this a long, long time ago.