emacs: eval-after-load instead of setopt for grep-command
Not sure how big of an effect this'd have on startup time, but I'm
leaning towards using the cheapest and simplest way to set each type of
variable (setq for global, setq-default for buffer-local,
eval-after-load that calls the setter for defcustom with setter), so I'm
avoiding setopt for now, which immediately loads the relevant package
and calls the appropriate setter. I guess a small downside with the
eval-after-load approach is that if you can't change the value after
init by re-evaluating the form.