vim: improve yanking for pipe and eval functions
For the operator-pending eval mappings, I doubt it ever makes sense to
take parts of multiple lines. So if multiple lines are selected, treat
it linewise. Try to simplify the logic by boiling it down to just two
dimensions: visual and linewise.

Use @" instead of @@, since that's how people usually use the unnamed
register. You only know about @@ if you've read the help for
:let-register recently.