statusbar: actually support spaces for tab completion
If we give `-o filenames` in the spec, then we can print the completions
unquoted and readline will quote them when it prints them. But if we do
that, it means that there's a mismatch between the unquoted candidates
we print, and the quoted ones that readline is adding to the command
line. For example, if the two candidates are 'two words' and 'two other
words', then after one TAB, cword will be 'two\ ', which won't match the
candidates in a construction like `[[ "$candidate" == "$cword"* ]]`.
compgen can deal with this for us, though, so it seems best to use it.
And we do need that backslash to be inserted for us, or else on the next
TAB we'll be trying to complete an empty word.