vim: add CopyAsBookmarklet
Also, disable the old temporary <leader>mb mapping but leave it as an
example for how NormalSurround was intended to be used.

Seems better to give the line numbers to CopyAsBookmarklet() as explicit
arguments instead of the using `range` function modifier, since running
`:%call <func>()` moves the cursor to the beginning of the range, and
doesn't even add the previous cursor position to the jumplist.