vim: add MungeAlone()
MungAlone() filters a range of lines using a Vim "munge" function,
similar to how :'<,'>!<cmd> would filter lines with a shell command (see
:h :range!).

The alternatives are awkward since you'd need to prefix most commands in
your munge function with a range. You'd probably want to use the `range`
modifier on the munge function, but you can't use variables in range
expressions without using :exe, so instead you'd probably want to to
save/restore some marks, and at that point you might have more
boilerplate than business logic. Most of the time it'd probably be fine
to compromise correctness a bit and select the given range as visual
mode (since it was most likely selected with visual mode, anyway) and
use the visual mode marks, but that's still harder to write/read and
more error prone than being able to assume the text is alone in a
buffer.