vim: simplify my image viewer bindings
Factoring out GetSelection() from PipeOp() makes the image viewer
bindings much easier to understand, and easier to run the viewer
command in the background.

Seems like starting a shell command in the background by suffixing it
with '&' is a fine way to have it run async from vim, assuming you don't
want to communicate with it further.