vim: let MakeX override the compiler, too
errorformats can be quite long, so we want to be able to pick them up
from the runtime files by setting the current "compiler", and also
override the command the compiler is called and do this conveniently
enough that we can do a one-liner mapping for a compiler without needing
to write a function for each one.

So, instead of doing something like:

    compiler jest
    let &l:makeprg = 'yarn jest'
    make!

We can instead do:

    MakeX(#{compiler: 'jest', makeprg: 'yarn jest'})

which is a lot easier to read in a key mapping.

oh, and MakeX() also undoes its changes, so that you can still set
makeprg and use :make normally, in addition to having keybindings for
other compilers.