vim: redo indentation settings
I can't think of when I'd want sw and sts to be different, so if sts=-1,
then I should hopefully never need to change it, so remove it.

Also, for files with et, tabs are an anomaly and it's not obvious how
long they were intended to be; if the indentation is a mix of tabs and
spaces, then 8-char tabs were often being used. So keep ts=8 except for
filetypes like go where it's common to use real tabs for indentation and
ts should equal sw. So most of the time sw is the only setting we need
to change.