vim: add JoinNoWhitespace again
I previously deleted it, but I actually do use it sometimes, though
probably not enough to need a binding. When using fluent interfaces,
JoinNoWhitespace is useful for coalescing a statement that has been
split across lines back onto one line.

This is a reimplementation in a more compact style, but I'm not aware of
anything being wrong with the previous implementation.