vim: add console.log() iabbrev
In the past I've used the initial letters of object for printf-debugging
abbreviations like this, but "cl" is a pretty common character sequence
and so I want to try out an abbreviation system that's universal across
languages and unlikely to conflict with anything. So, I'm thinking of
having 'z' as a leader, and the 'p' for 'print', and we can do 'pf' for
print-formatted, etc.