bash: bind kill-region to C-x C-k instead of C-w
Firstly, unix-rubout is pretty sweet, and I haven't used kill-region
since trying to add a binding for it. Also, apparently bash/readline
were interpreting "C-w" as $'C\000', which was overriding the default
`"C": self-insert` binding. Finally, bash/readline automatically bind
keys depending on your terminal settings, so if werase=^W in `stty`,
then any C-w binding in bashrc or inputrc will be overwritten anyway.
Seems best to bind kill-region in bash to something weird, especially
since in Emacs I can bind C-w to a function that does kill-region if
region is active or deletes the previous word if not.