vim: add node-test compiler
My impression from reading `:h errorformat` is that the quickfix list is
about parsing out locations and error messages, not displaying an entire
log file. If a line isn't matched it's corresponding entry is marked as
invalid, prefixed with `||`, and skipped when using the :cn and :cp
commands, but only if there are some valid entries. Sometimes it's nice
to have the full output, but other times it's not, so lately I'm
thinking that if I'm running tests where I need to look at a lot of
diagnostics, I'll either look back in my terminal's scrollback, or tee
the build output to a file and open a buffer for that, likely doing
`:setlocal autoread`. This compiler plugin follows that philosophy and
hides lines that aren't part of a quickfix entry.