bashrc: rewrite prompt
Using \x01 and \x02 instead of \[ \] to surround non-printing sequences
makes it way easier to write bash prompts, so I'm adding some fancier
features for my current situation, like abbreviating the working dir
based on an array of favourite dirs, and more concisely displaying
whether the activated python virtualenv is for the current directory.

My bash_completion script removes all pre-existing completions, so it
needs to be moved up in bashrc so that I can give completion specs for
functions.

Different distros put git-sh-prompt in different places, so vendor it in
~/.config. I do like having the sigils in my prompt. Sometimes checking
repo status takes a long time, so I've added an option, prompt_git, to
disable it temporarily.

If I have a virtualenv activated it's kind of awkward to apply bashrc
changes, so I've updated dedupe_path to put dirs matching *venv* first.
It's not bulletproof, but it's been very convenient so far.