fzf.vim: make swapchoice automatically when splitting
If a swap file already exists when opening a file, E325 is thrown, and
if you're trying to open the file in a split or new tab while also using
try/catch, the window creation gets cancelled during the error-handling,
leaving you with a buffer being created for the file but no window.

See https://github.com/junegunn/fzf.vim/issues/1514

Defining a SwapExists autocmd can avoid this, though, so use it whenever
opening more than 1 file or opening in a split/tab, and choose to open
the file as read-only, which is what I almost always want to do, anyway.
You can use `:e!` to choose a different option later.

https://github.com/Emilv2/siefe.vim gets around this by defining their
own alternative prompt for choosing swapchoice, but that doesn't seem
necessary for my purposes.