vim: use more semantic abbreviations for code
Instead of language-specific abbreviations like sop ->
System.out.println, use ;pl for "Print Line" across various languages.
It seems like an ending delimiter is the safest start token for these,
and semicolon is the most convenient ending delimiter to press; `]`
might be less likely to put keyword characters after, but it's a stretch
for my pinky.