Fix how GitHubUrl handles paths
I think I was giving a path relative to the repo root to `git ls-files`,
but was cd'ing via -C to the file's dir, and so ls-files wasn't printing
anything. I could give an absolute path to ls-files instead, but it
seems nice to explicitly check if the file is in a git repo.